Move the scratch-directory suite to its own file
The two tests above pushed NonInjectedNondeterminismVisitorTests.swift past the
500-line limit. The scratch-directory suite is self-contained and was the
natural cut.

No test changed; the file boundary did.
